# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Majed Sultan Al-Subaie Law Firm ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user with tenant theme and branding
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          description: Unauthorized
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CONTRACTS ──────────────────────────────────────────────────────────────
...
  /hearings:
    get:
      operationId: listHearings
      tags: [hearings]
      summary: List hearings
      parameters:
        - name: caseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hearing"
    post:
      operationId: createHearing
      tags: [hearings]
      summary: Create a hearing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HearingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hearing"
  /hearings/{id}:
    get:
      operationId: getHearing
      tags: [hearings]
      summary: Get hearing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hearing"
        "404":
          description: Not found
    patch:
      operationId: updateHearing
      tags: [hearings]
      summary: Update hearing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HearingUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hearing"
    delete:
      operationId: deleteHearing
      tags: [hearings]
      summary: Delete hearing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── EXECUTIONS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token, theme, branding]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
        theme:
          $ref: "#/components/schemas/TenantTheme"
        branding:
          $ref: "#/components/schemas/TenantBranding"
    SessionResponse:
      type: object
      required: [user, theme, branding]
      description: The authenticated user's profile plus their firm's theme and identity.
      properties:
        user:
          $ref: "#/components/schemas/User"
        theme:
          $ref: "#/components/schemas/TenantTheme"
        branding:
          $ref: "#/components/schemas/TenantBranding"
    TenantBranding:
      type: object
      required: [name, logoUrl]
      description: The authenticated firm's identity — display name and logo (null when the firm has no uploaded logo).
      properties:
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
    TenantTheme:
      type: object
      required: [primaryColor, secondaryColor]
      description: The authenticated firm's brand colors (hex), null when unset.
      properties:
        primaryColor:
          type: ["string", "null"]
        secondaryColor:
          type: ["string", "null"]
    User:
      type: object
      required: [id, email, role, tenantId, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [SYSTEM_MANAGER, TECHNICIAN]
        tenantId:
          type: integer
        createdAt:
          type: string

    # ─── DASHBOARD ────────────────────────────────────────────────────────────
...
    Client:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        nationalId:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        agencyNumber:
          type: ["string", "null"]
        agencyEndDate:
          type: ["string", "null"]
          format: date
        agencySource:
          type: ["string", "null"]
          enum: [خدمات الموثقين, الخدمات الالكترونية, null]
        createdAt:
          type: string
    ClientDetail:
      type: object
      required: [id, name, createdAt, cases, contracts]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        nationalId:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        agencyNumber:
          type: ["string", "null"]
        agencyEndDate:
          type: ["string", "null"]
          format: date
        agencySource:
          type: ["string", "null"]
          enum: [خدمات الموثقين, الخدمات الالكترونية, null]
        createdAt:
          type: string
        cases:
          type: array
          items:
            $ref: "#/components/schemas/Case"
        contracts:
          type: array
          items:
            $ref: "#/components/schemas/Contract"
    ClientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        nationalId:
          type: string
        address:
          type: string
        notes:
          type: string
        agencyNumber:
          type: string
        agencyEndDate:
          type: string
          format: date
        agencySource:
          type: string
          enum: [خدمات الموثقين, الخدمات الالكترونية]
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        nationalId:
          type: string
        address:
          type: string
        notes:
          type: string
        agencyNumber:
          type: ["string", "null"]
        agencyEndDate:
          type: ["string", "null"]
          format: date
        agencySource:
          type: ["string", "null"]
          enum: [خدمات الموثقين, الخدمات الالكترونية, null]

    # ─── CONTRACT ─────────────────────────────────────────────────────────────
    Contract:
      type: object
      required: [id, clientId, serviceType, hijriDate, isSigned, customClauses, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        serviceType:
          type: string
          enum: [FULL_REP, PARTIAL_REP, OBJECTION, CASSATION_REQUEST, CONTRACT_DRAFTING, CONTRACT_REVIEW, LEGAL_CONTRACT_CREATION, CONSULTATION]
        hijriDate:
          type: string
        gregorianDate:
          type: ["string", "null"]
        clientNationalId:
          type: ["string", "null"]
        clientAddress:
          type: ["string", "null"]
        clientPhone:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
        caseNumber:
          type: ["string", "null"]
        courtName:
          type: ["string", "null"]
        caseSubject:
          type: ["string", "null"]
        representationScope:
          type: ["string", "null"]
        preamble:
          type: ["string", "null"]
        fees:
          type: ["number", "null"]
        feeInstallments:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/FeeInstallment"
        isSigned:
          type: boolean
        customClauses:
          type: array
          items:
            type: string
        pdfUrl:
          type: ["string", "null"]
        signedPdfUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    FeeInstallment:
      type: object
      required: [description, amount, refundable]
      properties:
        description:
          type: string
        amount:
          type: number
        refundable:
          type: boolean
...
    Case:
      type: object
      required: [id, clientId, status, outcome, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        jurisdiction:
          type: ["string", "null"]
        clientRole:
          type: ["string", "null"]
          enum: [PLAINTIFF, DEFENDANT, null]
        opponentName:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        caseNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [UNDER_REVIEW, APPEAL, EXECUTION, CLOSED]
        outcome:
          type: string
          enum: [WON, LOST, PENDING]
        deletedAt:
          type: ["string", "null"]
        deletedByName:
          type: ["string", "null"]
        deletedByRole:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Hearing:
      type: object
      required: [id, caseId, hijriDate, utcDate, requiresLawsuitEditing, effectiveStatus, createdAt]
      properties:
        id:
          type: integer
        caseId:
          type: integer
        hijriDate:
          type: string
        utcDate:
          type: string
        status:
          type: ["string", "null"]
          enum: [UPCOMING, ENDED, CANCELLED, null]
        effectiveStatus:
          type: string
          enum: [UPCOMING, ENDED, CANCELLED]
        attendance:
          type: ["string", "null"]
        transcriptUrl:
          type: ["string", "null"]
        hearingReport:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        requiresLawsuitEditing:
          type: boolean
        requiresReplyPrep:
          type: boolean
        sessionLink:
          type: ["string", "null"]
        alertSent48h:
          type: boolean
        postHearingLocked:
          type: boolean
        createdAt:
          type: string
    HearingInput:
      type: object
      required: [caseId, hijriDate, utcDate]
      properties:
        caseId:
          type: integer
        hijriDate:
          type: string
        utcDate:
          type: string
        attendance:
          type: string
        transcriptUrl:
          type: string
        hearingReport:
          type: string
        notes:
          type: string
        requiresLawsuitEditing:
          type: boolean
        requiresReplyPrep:
          type: boolean
        sessionLink:
          type: string
    HearingUpdate:
      type: object
      properties:
        status:
          type: ["string", "null"]
          enum: [UPCOMING, ENDED, CANCELLED, null]
        hijriDate:
          type: string
        utcDate:
          type: string
        attendance:
          type: string
        transcriptUrl:
          type: string
        hearingReport:
          type: string
        notes:
          type: string
        requiresLawsuitEditing:
          type: boolean
        requiresReplyPrep:
          type: boolean
        sessionLink:
          type: string

    # ─── EXECUTION ────────────────────────────────────────────────────────────
...